# protocol/user/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Carry User API
  version: 0.1.0
servers:
  - url: /
security:
  - memberToken: []
  - browserSession: []
paths:
  /v1/browser/sessions:
    post:
      operationId: createBrowserSession
      security:
        - memberToken: []
      responses:
        "204":
          description: Browser session created
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "401":
          $ref: "#/components/responses/Error"
  /v1/browser/sessions/current:
    delete:
      operationId: revokeCurrentBrowserSession
      security:
        - browserSession: []
      responses:
        "204":
          description: Browser session revoked
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "401":
          $ref: "#/components/responses/Error"
  /v1/me:
    get:
      operationId: loadCurrentMember
      responses:
        "200":
          description: Current member and Spaces
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Error"
  /v1/machines/enroll:
    post:
      operationId: enrollMachine
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [space_id, display_name, public_key]
              properties:
                space_id:
                  type: string
                  format: uuid
                display_name:
                  type: string
                  minLength: 1
                public_key:
                  type: string
                  format: byte
      responses:
        "201":
          description: Machine enrolled or idempotently replayed
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineEnrollment"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/machines/revoke:
    post:
      operationId: revokeMachine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [space_id, machine_id]
              properties:
                space_id:
                  type: string
                  format: uuid
                machine_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Machine revoked
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status:
                    type: string
                    const: revoked
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/spaces/{spaceID}/conversation/messages:
    parameters:
      - $ref: "#/components/parameters/SpaceID"
    get:
      operationId: listConversationMessages
      parameters:
        - $ref: "#/components/parameters/BeforeConversationMessage"
        - $ref: "#/components/parameters/AfterConversationMessage"
      responses:
        "200":
          description: Newest or cursor-relative private Conversation page
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [messages]
                properties:
                  messages:
                    type: array
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/ConversationMessage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      operationId: sendConversationMessage
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  description: After trimming, 1–16384 UTF-8 bytes
      responses:
        "200":
          description: Accepted or idempotently replayed private member message
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationMessage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/spaces/{spaceID}/works:
    parameters:
      - $ref: "#/components/parameters/SpaceID"
    get:
      operationId: listWorks
      parameters:
        - $ref: "#/components/parameters/BeforeWork"
      responses:
        "200":
          description: >-
            Newest or cursor-relative Work summaries visible in the Space
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [works, has_earlier_works]
                properties:
                  works:
                    type: array
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/WorkSummary"
                  has_earlier_works:
                    type: boolean
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      operationId: createWork
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [goal]
              properties:
                goal:
                  type: string
                  minLength: 1
                  description: After trimming, 1–2000 UTF-8 bytes
      responses:
        "200":
          description: Created or idempotently replayed Work
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/spaces/{spaceID}/works/{workID}:
    parameters:
      - $ref: "#/components/parameters/SpaceID"
      - $ref: "#/components/parameters/WorkID"
    get:
      operationId: loadWork
      parameters:
        - $ref: "#/components/parameters/BeforeWorkMessage"
      responses:
        "200":
          description: >-
            Work and its newest or cursor-relative ordered member message page
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [work, messages, has_earlier_messages]
                properties:
                  work:
                    $ref: "#/components/schemas/Work"
                  messages:
                    type: array
                    maxItems: 50
                    description: >-
                      Ordered complete messages with at most 262144 total
                      UTF-8 text bytes
                    items:
                      $ref: "#/components/schemas/WorkMessage"
                  has_earlier_messages:
                    type: boolean
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/spaces/{spaceID}/works/{workID}/messages:
    parameters:
      - $ref: "#/components/parameters/SpaceID"
      - $ref: "#/components/parameters/WorkID"
    post:
      operationId: appendWorkMessage
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  description: Non-blank and at most 61440 UTF-8 bytes
      responses:
        "200":
          description: Appended or idempotently replayed Work Message
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMessage"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/spaces/{spaceID}/works/{workID}/retry:
    parameters:
      - $ref: "#/components/parameters/SpaceID"
      - $ref: "#/components/parameters/WorkID"
    post:
      operationId: retryWork
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Work retry requested or idempotently replayed
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
components:
  headers:
    NoStore:
      description: Authenticated responses must not be stored by any cache
      schema:
        type: string
        const: no-store
  securitySchemes:
    memberToken:
      type: http
      scheme: bearer
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-carry_session
  parameters:
    SpaceID:
      name: spaceID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WorkID:
      name: workID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
    BeforeConversationMessage:
      name: before
      in: query
      required: false
      schema:
        type: string
        format: uuid
    AfterConversationMessage:
      name: after
      in: query
      required: false
      schema:
        type: string
        format: uuid
    BeforeWork:
      name: before
      in: query
      required: false
      schema:
        type: string
        format: uuid
    BeforeWorkMessage:
      name: before
      in: query
      required: false
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Request rejected
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
    APIError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    Membership:
      type: object
      additionalProperties: false
      required: [space_id, name, can_enroll_machines]
      properties:
        space_id:
          type: string
          format: uuid
        name:
          type: string
        can_enroll_machines:
          type: boolean
    Member:
      type: object
      additionalProperties: false
      required: [user_id, display_name, spaces]
      properties:
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        spaces:
          type: array
          items:
            $ref: "#/components/schemas/Membership"
    MachineEnrollment:
      type: object
      additionalProperties: false
      required: [machine_id, space_id, certificate_pem]
      properties:
        machine_id:
          type: string
          format: uuid
        space_id:
          type: string
          format: uuid
        certificate_pem:
          type: string
    ConversationMessage:
      type: object
      additionalProperties: false
      required: [message_id, author, text, created_at]
      properties:
        message_id:
          type: string
          format: uuid
        author:
          type: string
          enum: [member, carry]
        text:
          type: string
        request_id:
          type: string
          description: Present only on the authenticated member's own input
        created_work_id:
          type: string
          format: uuid
          description: Present only on a private Carry reply that created Work
        created_at:
          type: string
          format: date-time
    Work:
      type: object
      additionalProperties: false
      required:
        - work_id
        - space_id
        - goal
        - lifecycle
        - owner_user_id
        - owner_display_name
        - creator_user_id
        - creator_display_name
        - understanding
        - next_step
        - has_unapplied_input
        - needs_retry
        - created_at
      properties:
        work_id:
          type: string
          format: uuid
        space_id:
          type: string
          format: uuid
        goal:
          type: string
        lifecycle:
          type: string
          enum: [open]
        owner_user_id:
          type: string
          format: uuid
        owner_display_name:
          type: string
        creator_user_id:
          type: string
          format: uuid
        creator_display_name:
          type: string
        understanding:
          type: string
        next_step:
          type: string
        has_unapplied_input:
          type: boolean
        needs_retry:
          type: boolean
          description: An active member must explicitly request a fresh attempt
        created_at:
          type: string
          format: date-time
    WorkSummary:
      type: object
      additionalProperties: false
      required:
        - work_id
        - space_id
        - goal
        - lifecycle
        - owner_user_id
        - owner_display_name
        - creator_user_id
        - creator_display_name
        - has_unapplied_input
        - needs_retry
        - created_at
      properties:
        work_id:
          type: string
          format: uuid
        space_id:
          type: string
          format: uuid
        goal:
          type: string
        lifecycle:
          type: string
          enum: [open]
        owner_user_id:
          type: string
          format: uuid
        owner_display_name:
          type: string
        creator_user_id:
          type: string
          format: uuid
        creator_display_name:
          type: string
        has_unapplied_input:
          type: boolean
        needs_retry:
          type: boolean
          description: An active member must explicitly request a fresh attempt
        created_at:
          type: string
          format: date-time
    WorkMessage:
      type: object
      additionalProperties: false
      required:
        - message_id
        - work_id
        - author_user_id
        - author_display_name
        - text
        - created_at
      properties:
        message_id:
          type: string
          format: uuid
        work_id:
          type: string
          format: uuid
        author_user_id:
          type: string
          format: uuid
        author_display_name:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time